requirements.txt
Django==1.9
Markdown==2.6.5
Pillow==3.0.0
PyJWT==1.4.0
PyYAML==3.11
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
djoser==0.4.0
drf-extensions==0.2.8
